- :sparkles: (External, Process) extract to base class the external process handling
- Merge branch 'feature/external-process' of https://dev.azure.com/elvexoft/Nexai/_git/Elvex.Toolbox into feature/external-process
- clean up
- :sparkles: (FileSystem, Service) Add folder delete

+++ src/Elvex.Toolbox/Services/FileSystemHandler.cs
@@ -154,8 +154,23 @@ namespace Elvex.Toolbox.Services
         public ValueTask<bool> Delete(Uri file)
         {
             var exists = Exists(file);
-            File.Delete(file.OriginalString);
+            if (exists)
+                File.Delete(file.OriginalString);
             return ValueTask.FromResult(exists);
+        }
+
+        /// <summary>
+        /// Deletes the specified folder
+        /// </summary>
+        public ValueTask<bool> DeleteFolder(Uri folder, bool recurive = true)
+        {
+            if (Directory.Exists(folder.LocalPath))
+            {
+                Directory.Delete(folder.LocalPath, recurive);
+                return ValueTask.FromResult(true);
+            }
+
+            return ValueTask.FromResult(false);
         }
 
         /// <inheritdoc />
